% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYplot.R
\docType{package}
\name{SYplot}
\alias{SYplot}
\alias{SYplot-package}
\title{SYplot: A package for}
\description{
The SYplot package provides
}
\section{Foo functions}{

The foo functions ...
}

\author{
\strong{Maintainer}: Joost Donkers \email{joost@protonmail.com}

}
