% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syr_admin2sf.R
\docType{data}
\name{syr_adm2sf}
\alias{syr_adm2sf}
\title{Syria admin level 2 boudaries}
\format{A data frame with 60 rows and 18 variables:
\describe{
  \item{OBJECTID}{unique poligon number}
  \item{NAME_2}{Name of Mohafaza}
  ...
}}
\source{
\url{https://data.humdata.org/dataset/syrian-arab-republic-administrative-boundaries-populated-places}
}
\usage{
syr_adm2sf
}
\description{
boundaries of nahyas in SF format
}
\keyword{datasets}
